[fix] a & b precision

+++ src/draw.hpp
@@ -15,4 +15,4 @@ void draw_init( SDL_Renderer * render );
 void set_coloru( Uint32 color );
 void set_color3u( Uint8 red, Uint8 green, Uint8 blue );
 void set_color4u( Uint8 red, Uint8 green, Uint8 blue, Uint8 alpha );
-void draw_ellipse( unsigned int x, unsigned int y, unsigned int a, unsigned int b );
+void draw_ellipse( unsigned int x, unsigned int y, float a, float b );
